Split service and component
